Upgrade Gradle to 8.12 and fix iOS framework linking
This commit upgrades Gradle to version 8.12 and fixes framework linking for iOS arm64 and simulator arm64 targets.

The framework linking fix ensures that `linkerOpts` are applied to the correct binaries in the iOS targets.

Additionally, a .kotlin directory is added to the .gitignore file.

I submit my source code and allow Tigase to release the code as a part of Tigase's own software package under any license, either an open source or closed source.